+++ src/descriptors/element_size.js
@@ -1,0 +1,53 @@
+// Copyright (c) 2014 Titanium I.T. LLC. All rights reserved. For license, see "README" or "LICENSE" file.
+"use strict";
+
+var ensure = require("../util/ensure.js");
+var Descriptor = require("./descriptor.js");
+var Size = require("../values/size.js");
+
+var X_DIMENSION = "x";
+var Y_DIMENSION = "y";
+
+var Me = module.exports = function ElementSize(dimension, element) {
+	// TODO: circular dependency prevents ensure.signature
+	ensure.that(dimension === X_DIMENSION || dimension === Y_DIMENSION, "Unrecognized dimension: " + dimension);
+
+	this._dimension = dimension;
+	this._element = element;
+};
+Descriptor.extend(Me);
+
+Me.x = function x(element) {
+	return new Me(X_DIMENSION, element);
+};
+
+Me.y = function y(element) {
+	return new Me(Y_DIMENSION, element);
+};
+
+Me.prototype.value = function value() {
+	ensure.signature(arguments, []);
+
+	var position = this._element.getRawPosition();
+	var result = (this._dimension === X_DIMENSION) ? position.width : position.height;
+
+	return new Size(result);
+};
+
+Me.prototype.convert = function convert(arg) {
+	ensure.signature(arguments, [ [Number, Descriptor] ]);
+	if (typeof arg !== "number") return arg;
+
+	return new Size(arg);
+};
+
+Me.prototype.describeMatch = function describeMatch() {
+	return "match " + this.toString() + " (" + this.value() + ")";
+};
+
+Me.prototype.toString = function toString() {
+	ensure.signature(arguments, []);
+
+	var desc = (this._dimension === X_DIMENSION) ? "width" : "height";
+	return desc + " of element '" + this._element.description() + "'";
+};